Hi Really Sorry for such a large commit. I had some issues with Git as this part of my project was made on my work laptop and wasnt able to merge the files. I will comment out all of the code for you so that you are able to see what each part does. Hope this helps.
Please see a break down of the changes below:

Dependencies have been added to the build gradle file for screenshott.
There are java files associated with this which have also been added.

Added Image files to my drawable folder for the layout.
Added sound files for the button click into res - raw folder.

Created 2 new java class files for my recycler view and adapter.
Created layout for the sample card view layout. I decided to keep this part nice and simple.

Initially I tried using volley with Network Image View to parse the JSON Data however I felt this was a little sluggish as volley is a little out of date now.
In order to better the end user experience I decided to use Picasso to handle the Image requests but still used volley to parse the data.
This kept my code a lot cleaner.

I then created another blank activity with an image view to hold the images pulled accross via the ONCLick method from JSON.

I then updated permissions in the manifest file to allow the app write access to External storage.

I tried a few different ways to achieve adding the banners to the images via a web server however I felt that It would be a simpler solution to overlap images in the layout using button clicks.
In order to save the images I found a way of taking a screen shot of the Layout its self rather than calling the Image files directly.

This also added to the user experience as it now feels like they are actually taking a photo and saving it like they would if they used a camera. Which is what my apps idea is based on :)

I did a lot of research to on stack overflow on how to achive this progammaticly but a lot of the solutions required root access. I wasnt able to impliment this myself in a timely manner, so I opted to use a great 3rd party lib to do so.

upon taking the screenhot the user also has the option to save the file to their storage. I added a URI Intent to open the user gallery. This way they are able to click on the image and edit/share with their phones native software. Its not a perfect solution to my issue but it most definatly works.

I would love to spend a bit more time with this in the future to learn best practices and find out what would be the best way to go about this.

Im sure I will pick your brains on this during the interview section :)